Host clock sync: minor fixes for non-ancient VMware WS versions.
- All involved data types are ulong, not int32.
- VMW_BACK_GET_HOST_TIME doesn't returns timezone offset anymore.
- Reduce polling interval to once per minute (same value used
  by default on the official "open_vmware_tools").

Note: "VMW_BACK_GET_HOST_TIME" (=23) is deprecated.

We should switch to using the suggested "*_GETTIMEFULL" (=46)
command instead.

+++ shared_sources/VMWBackdoor.h
@@ -21,7 +21,7 @@ public:
 	void		GetCursorStatus(uint16& status, uint16& to_read);
 	status_t	GetHostClipboard(char** text, size_t *text_length);
 	status_t	SetHostClipboard(char* text, size_t length);
-	int32		GetHostClock();
+	ulong		GetHostClock();
 };
 
 #endif
